pyproject.toml
[project]
name = "agent_cv"
version = "0.1.0"
description = "Vision + RAG Agent"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain-core",
    "langchain-ollama",
    "langchain-teddynote",
    "langgraph",
    "langsmith",
    "pydantic",
    "requests",
    "python-dotenv",
    "langserve[all] (>=0.3.3,<0.4.0)",
    "streamlit (>=1.53.1,<2.0.0)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
